pyproject.toml
[tool.poetry]
name = "command-ghostwriter"
version = "0.1.0-alpha.0"
description = ""
authors = ["tvna <tsubasa.nagano@icloud.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.34.0"
jinja2 = "^3.1.4"
pyyaml = "^6.0.1"
micropip = "^0.6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
flake8-pyproject = "^1.2.3"
flake8-pytest-style = "^2.0.0"
flake8-builtins = "^2.5.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
types-toml = "^0.10.8.20240310"
pre-commit = "^3.7.1"
lizard = "^1.17.10"
cohesion = "^1.1.0"
pep8-naming = "^0.13.3"
poetry-types = "^0.5.1"
types-pyyaml = "^6.0.12.20240311"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests",]
pythonpath = "."
addopts = [
    "--import-mode=importlib",
]

[tool.isort]
line_length = 140

[tool.black]
target-version = ['py312']
line-length = 140

[tool.mypy]
python_version = "3.12"
no_strict_optional = true
ignore_missing_imports = false
check_untyped_defs = true